# Optimize: reduce allocs