Use fixed size slice to reduce allocs.

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkSimple/ASCII-8       324           294           -9.09%
BenchmarkSimple/French-8      620           584           -5.71%
BenchmarkSimple/Nordic-8      1187          1188          +0.08%
BenchmarkSimple/Tibetan-8     1044          1016          -2.68%

benchmark                     old allocs     new allocs     delta
BenchmarkSimple/ASCII-8       1              0              -100.00%
BenchmarkSimple/French-8      1              0              -100.00%
BenchmarkSimple/Nordic-8      1              0              -100.00%
BenchmarkSimple/Tibetan-8     1              0              -100.00%

benchmark                     old bytes     new bytes     delta
BenchmarkSimple/ASCII-8       24            0             -100.00%
BenchmarkSimple/French-8      32            0             -100.00%
BenchmarkSimple/Nordic-8      48            0             -100.00%
BenchmarkSimple/Tibetan-8     48            0             -100.00%

```