@agnivade 2021-05-01
A neat trick! I would like to refactor this code a bit more and add some more tests. Would you mind if I push another commit to your fork?

I'll rebase the changes so that the commit history is preserved of course. Let me know if that's okay.